Actions contain a seperation between unit and unit_ref as well as target and target_ref
unit_ref and target_ref may contain out-of-date references, while unit and target are up to date.

This distinction is important, because they may refer to different underlying objects. This is not important if you are merely reading from the object, but if the object is changed, it is important.