In this pull request I will make all the unit dictionaries (traits, states, abilities and effects) contain a list of two items: level and value. Since at least effects needed to have both, it was too annoying to make special exceptions for that one. Initially I wanted to make it a namedtuple or class, so that I could write trait.level instead of trait[1]. But according some time tests I ran, this would increase the running time significantly:
class 0.115000009537
namedtuple 0.131000041962
list 0.0390000343323
int 0.0149998664856
(10000 instantiated)

So I will start with a list, and see if the time required for this is a bottleneck. If it isn't, I will change it to a class later.

The universal_tests still look the same, except that if a unit is targeted by an effect, I will write for example:
"poisoned": [2,2] instead of
"poisoned_II": 2